주문 관련 기능 구현 완료

+++ karma/src/main/java/com/ecyce/karma/domain/order/dto/OrderCreateRequestDto.java
@@ -20,5 +20,5 @@ public class OrderCreateRequestDto {
     private String request;
 
     @NotNull
-    private Long orderCount;
+    private Integer orderCount;
 }
